POM_09_Element Displayed/Undisplayed _ Overide Timeout

+++ actions/commons/Constants.java
@@ -3,6 +3,6 @@ package commons;
 public class Constants {
 	public static final String DEV_APP_URL = "http://demo.guru99.com/v4";
 	public static final String TEST_APP_URL = "http://demo.guru99.com/v1";
-	public static final int SHORT_TIMEOUT = 5;
-	public static final int LONG_TIMEOUT = 30;
+	public static final long SHORT_TIMEOUT = 5;
+	public static final long LONG_TIMEOUT = 30;
 }
